move all header to a .hpp format (#19)
