+++ include/dynamic_graph_manager/exception/exception-abstract.hpp
@@ -1,5 +1,5 @@
 /**
- * @file exception-abstract.hh
+ * @file exception-abstract.hpp
  * @author Maximilien Naveau (maximilien.naveau@gmail.com)
  * @license License BSD-3-Clause
  * @copyright Copyright (c) 2019, New York University and Max Planck
