+++ include/dynamic_graph_manager/exception/exception-task.hpp
@@ -1,5 +1,5 @@
 /**
- * @file exception-tools.hh
+ * @file exception-task.hpp
  * @author Maximilien Naveau (maximilien.naveau@gmail.com)
  * @license License BSD-3-Clause
  * @copyright Copyright (c) 2019, New York University and Max Planck
@@ -7,59 +7,51 @@
  * @date 2019-05-22
  */
 
-#ifndef TOOLS_EXCEPTION_HH
-#define TOOLS_EXCEPTION_HH
+#ifndef EXCEPTION_TASK_HH
+#define EXCEPTION_TASK_HH
 
 /* --------------------------------------------------------------------- */
 /* --- INCLUDE --------------------------------------------------------- */
 /* --------------------------------------------------------------------- */
 
-#include <dynamic_graph_manager/exception/exception-abstract.hh>
+#include <dynamic_graph_manager/exception/exception-abstract.hpp>
+
 /* --------------------------------------------------------------------- */
 /* --- CLASS ----------------------------------------------------------- */
 /* --------------------------------------------------------------------- */
 
 namespace dynamic_graph
 {
-/* \class ExceptionTools
- */
-class ExceptionTools : public ExceptionAbstract
+/** \brief ExceptionTask */
+class ExceptionTask : public ExceptionAbstract
 {
 public:
     enum ErrorCodeEnum
     {
-        GENERIC = ExceptionAbstract::TOOLS
-
-        ,
-        CORBA,
-        KALMAN_SIZE,
-        PY_SHELL_PTR
+        GENERIC = ExceptionAbstract::TASK,
+        EMPTY_LIST,
+        NON_ADEQUATE_FEATURES,
+        MATRIX_SIZE,
+        BOUND_TYPE,
+        PARSER_MULTI_BOUND
     };
 
     static const std::string EXCEPTION_NAME;
-    virtual const std::string& getExceptionName() const
+    virtual const std::string& getExceptionName(void) const
     {
         return EXCEPTION_NAME;
     }
 
-public:
-    ExceptionTools(const ExceptionTools::ErrorCodeEnum& errcode,
-                   const std::string& msg = "");
-    ExceptionTools(const ExceptionTools::ErrorCodeEnum& errcode,
-                   const std::string& msg,
-                   const char* format,
-                   ...);
-    virtual ~ExceptionTools(void) throw()
+    ExceptionTask(const ExceptionTask::ErrorCodeEnum& errcode,
+                  const std::string& msg = "");
+    ExceptionTask(const ExceptionTask::ErrorCodeEnum& errcode,
+                  const std::string& msg,
+                  const char* format,
+                  ...);
+    virtual ~ExceptionTask(void) throw()
     {
     }
 };
-
 }  // namespace dynamic_graph
 
-#endif /* #ifndef TOOLS_EXCEPTION_HH */
-
-/*
- * Local variables:
- * c-basic-offset: 2
- * End:
- */
+#endif /* #ifndef EXCEPTION_TASK_HH */
