+++ include/dynamic_graph_manager/exception/exception-tools.hpp
@@ -1,5 +1,5 @@
 /**
- * @file exception-feature.hh
+ * @file exception-tools.hpp
  * @author Maximilien Naveau (maximilien.naveau@gmail.com)
  * @license License BSD-3-Clause
  * @copyright Copyright (c) 2019, New York University and Max Planck
@@ -7,55 +7,56 @@
  * @date 2019-05-22
  */
 
-#ifndef EXCEPTION_FEATURE_HH
-#define EXCEPTION_FEATURE_HH
+#ifndef TOOLS_EXCEPTION_HH
+#define TOOLS_EXCEPTION_HH
 
 /* --------------------------------------------------------------------- */
 /* --- INCLUDE --------------------------------------------------------- */
 /* --------------------------------------------------------------------- */
 
-#include <dynamic_graph_manager/exception/exception-abstract.hh>
-
+#include <dynamic_graph_manager/exception/exception-abstract.hpp>
 /* --------------------------------------------------------------------- */
 /* --- CLASS ----------------------------------------------------------- */
 /* --------------------------------------------------------------------- */
 
 namespace dynamic_graph
 {
-/* \class ExceptionFeature
+/* \class ExceptionTools
  */
-class ExceptionFeature : public ExceptionAbstract
-
+class ExceptionTools : public ExceptionAbstract
 {
 public:
     enum ErrorCodeEnum
     {
-        GENERIC = ExceptionAbstract::FEATURE,
-        BAD_INIT,
-        UNCOMPATIBLE_SIZE
+        GENERIC = ExceptionAbstract::TOOLS
+
+        ,
+        CORBA,
+        KALMAN_SIZE,
+        PY_SHELL_PTR
     };
 
     static const std::string EXCEPTION_NAME;
-    virtual const std::string& getExceptionName(void) const
+    virtual const std::string& getExceptionName() const
     {
-        return ExceptionFeature::EXCEPTION_NAME;
+        return EXCEPTION_NAME;
     }
 
-    ExceptionFeature(const ExceptionFeature::ErrorCodeEnum& errcode,
-                     const std::string& msg = "");
-
-    ExceptionFeature(const ExceptionFeature::ErrorCodeEnum& errcode,
-                     const std::string& msg,
-                     const char* format,
-                     ...);
-
-    virtual ~ExceptionFeature(void) throw()
+public:
+    ExceptionTools(const ExceptionTools::ErrorCodeEnum& errcode,
+                   const std::string& msg = "");
+    ExceptionTools(const ExceptionTools::ErrorCodeEnum& errcode,
+                   const std::string& msg,
+                   const char* format,
+                   ...);
+    virtual ~ExceptionTools(void) throw()
     {
     }
 };
+
 }  // namespace dynamic_graph
 
-#endif /* #ifndef EXCEPTION_FEATURE_HH */
+#endif /* #ifndef TOOLS_EXCEPTION_HH */
 
 /*
  * Local variables:
