+++ include/dynamic_graph_manager/periodic-call.hpp
@@ -1,5 +1,5 @@
 ﻿/**
- * @file periodic-call.hh
+ * @file periodic-call.hpp
  * @author Maximilien Naveau (maximilien.naveau@gmail.com)
  * @license License BSD-3-Clause
  * @copyright Copyright (c) 2019, New York University and Max Planck
