+++ include/dynamic_graph_manager/ros_init.hpp
@@ -1,5 +1,5 @@
 /**
- * @file ros_init.hh
+ * @file ros_init.hpp
  * @author Maximilien Naveau (maximilien.naveau@gmail.com)
  * @license License BSD-3-Clause
  * @copyright Copyright (c) 2019, New York University and Max Planck
