+++ include/dynamic_graph_manager/ros_interpreter.hpp
@@ -1,5 +1,5 @@
 /**
- * @file ros_interpreter.hh
+ * @file ros_interpreter.hpp
  * @author Maximilien Naveau (maximilien.naveau@gmail.com)
  * @license License BSD-3-Clause
  * @copyright Copyright (c) 2019, New York University and Max Planck
