+++ include/ros_entities/matrix_geometry.hpp
@@ -1,5 +1,5 @@
 /**
- * @file matrix_geometry.hh
+ * @file matrix_geometry.hpp
  * @author Maximilien Naveau (maximilien.naveau@gmail.com)
  * @license License BSD-3-Clause
  * @copyright Copyright (c) 2019, New York University and Max Planck
