+++ src/ros_entities/ros_publish.cpp
@@ -24,10 +24,10 @@
 #include <dynamic-graph/factory.h>
 #include <dynamic-graph/linear-algebra.h>
 
-#include "dynamic_graph_manager/dynamic_graph_manager.hh"
-#include "dynamic_graph_manager/ros_init.hh"
+#include "dynamic_graph_manager/dynamic_graph_manager.hpp"
+#include "dynamic_graph_manager/ros_init.hpp"
 
-#include "ros_entities/ros_publish.hh"
+#include "ros_entities/ros_publish.hpp"
 
 namespace dynamic_graph
 {
